requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
dnspython==2.1.0
python-dotenv==0.19.1
expiringdict==1.2.1
Werkzeug==2.3.7
requests==2.26.0

pandas==2.1.1
pymongo==4.7.2
numpy==1.26.0